feat: refactor and add DialogProvider
