--- src/views/components/Settings.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-type Props = {
-    className?: string;
-};
-
-/**
- * Component to hold everything for the settings page
- * @param props className
- * @returns The content for the settings page
- */
-export default function Settings({ className }: Props): JSX.Element {
-    // TODO: Implement the settings page
-    return <div className={className}>this will be finished laterrrrrrr</div>;
-}
